UPSTREAM: seq_buf: Move the seq_buf code to lib/
The seq_buf functions are rather useful outside of tracing. Instead
of having it be dependent on CONFIG_TRACING, move the code into lib/
and allow other users to have access to it even when tracing is not
configured.

The seq_buf utility is similar to the seq_file utility, but instead of
writing sending data back up to userland, it writes it into a buffer
defined at seq_buf_init(). This allows us to send a descriptor around
that writes printf() formatted strings into it that can be retrieved
later.

It is currently used by the tracing facility for such things like trace
events to convert its binary saved data in the ring buffer into an
ASCII human readable context to be displayed in /sys/kernel/debug/trace.

It can also be used for doing NMI prints safely from NMI context into
the seq_buf and retrieved later and dumped to printk() safely. Doing
printk() from an NMI context is dangerous because an NMI can preempt
a current printk() and deadlock on it.

Link: http://lkml.kernel.org/p/20140619213952.058255809@goodmis.org

BUG=chrome-os-partner:51198
TEST=sudo -i;
     echo 10 > /proc/sys/kernel/hung_task_timeout_secs;
     echo HUNG_TASK > /sys/kernel/debug/provoke-crash/DIRECT &

Tested-by: Jiri Kosina <jkosina@suse.cz>
Acked-by: Jiri Kosina <jkosina@suse.cz>
Reviewed-by: Petr Mladek <pmladek@suse.cz>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
(cherry picked from commit 8d58e99af5980d444948720977b0976455885391
 git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git)
Signed-off-by: Harry Pan <harry.pan@intel.com>

Change-Id: I3c8f69f46b311f7c460e317c3d4ffac280120ca9
Reviewed-on: https://chromium-review.googlesource.com/337901
Commit-Ready: Harry Pan <harry.pan@intel.com>
Tested-by: Zhuo-hao Lee <zhuo-hao.lee@intel.com>
Reviewed-by: Benson Leung <bleung@chromium.org>