UPSTREAM: printk: Add per_cpu printk func to allow printk to be diverted
Being able to divert printk to call another function besides the normal
logging is useful for such things like NMI handling. If some functions
are to be called from NMI that does printk() it is possible to lock up
the box if the nmi handler triggers when another printk is happening.

One example of this use is to perform a stack trace on all CPUs via NMI.
But if the NMI is to do the printk() it can cause the system to lock up.
By allowing the printk to be diverted to another function that can safely
record the printk output and then print it when it in a safe context
then NMIs will be safe to call these functions like show_regs().

Link: http://lkml.kernel.org/p/20140619213952.209176403@goodmis.org

BUG=chrome-os-partner:51198
TEST=sudo -i;
     echo 10 > /proc/sys/kernel/hung_task_timeout_secs;
     echo HUNG_TASK > /sys/kernel/debug/provoke-crash/DIRECT &

Tested-by: Jiri Kosina <jkosina@suse.cz>
Acked-by: Jiri Kosina <jkosina@suse.cz>
Acked-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>
Reviewed-by: Petr Mladek <pmladek@suse.cz>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
(cherry picked from commit afdc34a3d3b823a12a93b822ee1efb566f884032
 git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git)
Signed-off-by: Harry Pan <harry.pan@intel.com>

Change-Id: I24771e665a6d9571f2dd9b3eed804103a2f9fb10
Reviewed-on: https://chromium-review.googlesource.com/337902
Commit-Ready: Harry Pan <harry.pan@intel.com>
Tested-by: Zhuo-hao Lee <zhuo-hao.lee@intel.com>
Reviewed-by: Benson Leung <bleung@chromium.org>