UPSTREAM: x86/nmi: Perform a safe NMI stack trace on all CPUs
When trigger_all_cpu_backtrace() is called on x86, it will trigger an
NMI on each CPU and call show_regs(). But this can lead to a hard lock
up if the NMI comes in on another printk().

In order to avoid this, when the NMI triggers, it switches the printk
routine for that CPU to call a NMI safe printk function that records the
printk in a per_cpu seq_buf descriptor. After all NMIs have finished
recording its data, the seq_bufs are printed in a safe context.

Link: http://lkml.kernel.org/p/20140619213952.360076309@goodmis.org
Link: http://lkml.kernel.org/r/20141115050605.055232587@goodmis.org

BUG=chrome-os-partner:51198
TEST=sudo -i;
     echo 10 > /proc/sys/kernel/hung_task_timeout_secs;
     echo HUNG_TASK > /sys/kernel/debug/provoke-crash/DIRECT &

Tested-by: Jiri Kosina <jkosina@suse.cz>
Acked-by: Jiri Kosina <jkosina@suse.cz>
Acked-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>
Reviewed-by: Petr Mladek <pmladek@suse.cz>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
(cherry picked from commit a9edc88093287183ac934be44f295f183b2c62dd
 git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git)
Signed-off-by: Harry Pan <harry.pan@intel.com>

Change-Id: Ieb8ce948fe2455e9d4b5a4ff49571cf8679a5f94
Reviewed-on: https://chromium-review.googlesource.com/337903
Commit-Ready: Harry Pan <harry.pan@intel.com>
Tested-by: Zhuo-hao Lee <zhuo-hao.lee@intel.com>
Reviewed-by: Benson Leung <bleung@chromium.org>