UPSTREAM: HID: logitech: allow the DJ device to request the unifying name
The names of the DJ devices are stored in the receiver. These names
can be retrieved through a HID++ command. However, the protocol says
that you have to ask the receiver for that, not the device iteself.

Introduce a special case in the DJ handling where a device can request
its unifying name, and when such a name is given, forward it also to
the corresponding device.

On the HID++ side, the receiver talks only HID++ 1.0, so we need to
implement this part of the protocol in the module.

BUG=chromium:617976
TEST=build

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Tested-by: Andrew de los Reyes <adlr@chromium.org>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit 33797820af98cde5c7cee00d00f0d8e255ea199f)
Signed-off-by: Adam Goode <agoode@chromium.org>

Change-Id: I5aad283262f8951a3e4bfae6498079f007476b60
Reviewed-on: https://chromium-review.googlesource.com/361106
Commit-Ready: Adam Goode <agoode@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>