UPSTREAM: tracing: Have seq_buf use full buffer
Currently seq_buf is full when all but one byte of the buffer is
filled. Change it so that the seq_buf is full when all of the
buffer is filled.

Some of the functions would fill the buffer completely and report
everything was fine. This was inconsistent with the max of size - 1.
Changing this to be max of size makes all functions consistent.

Link: http://lkml.kernel.org/r/20141104160222.502133196@goodmis.org
Link: http://lkml.kernel.org/r/20141114011412.811957882@goodmis.org

BUG=chrome-os-partner:51198
TEST=sudo -i;
     echo 10 > /proc/sys/kernel/hung_task_timeout_secs;
     echo HUNG_TASK > /sys/kernel/debug/provoke-crash/DIRECT &

Tested-by: Jiri Kosina <jkosina@suse.cz>
Acked-by: Jiri Kosina <jkosina@suse.cz>
Reviewed-by: Petr Mladek <pmladek@suse.cz>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
(cherry picked from commit 8cd709ae7658a7fd7f6630699e3229188c2591e4
 git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git)
Signed-off-by: Harry Pan <harry.pan@intel.com>

Change-Id: I38c95b1a849499a78f3d8d00c31702487ed0c1a6
Reviewed-on: https://chromium-review.googlesource.com/337898
Commit-Ready: Harry Pan <harry.pan@intel.com>
Tested-by: Zhuo-hao Lee <zhuo-hao.lee@intel.com>
Reviewed-by: Benson Leung <bleung@chromium.org>