UPSTREAM: dmaengine: Create a generic dma_slave_caps callback
dma_slave_caps is very important to the generic layers that might interact with
dmaengine, such as ASoC. Unfortunately, it has been added as yet another
dma_device callback, and most of the existing drivers haven't implemented it,
reducing its reliability.

Introduce a generic behaviour to implement this, that rely on both the split of
device_control to derive which functions are supported and on new variables to
be set in the dma_device structure.

These variables holds what used to be the capabilities, that were set
per-channel. However, this proved to be a bit overkill, since every driver
filling these so far were hardcoding it, disregarding which channel was
actually given.

Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>
Acked-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Vinod Koul <vinod.koul@intel.com>
(cherry picked from commit cb8cea513c80db1dfe2dce468d2d0772005bb9a1)

BUG=chrome-os-partner:49255
TEST=Serial and DMA are working with these changes

Change-Id: I12f0fd2072114b731e37140d15fcbc78a9cff653
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/332996
Commit-Ready: Dmitry Torokhov <dtor@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>