UPSTREAM: tracing: Clean up tracing_fill_pipe_page()
The function tracing_fill_pipe_page() logic is a little confusing with the
use of count saving the seq.len and reusing it.

Instead of subtracting a number that is calculated from the saved
value of the seq.len from seq.len, just save the seq.len at the start
and if we need to reset it, just assign it again.

When the seq_buf overflow is len == size + 1, the current logic will
break. Changing it to use a saved length for resetting back to the
original value is more robust and will work when we change the way
seq_buf sets the overflow.

Link: http://lkml.kernel.org/r/20141118161546.GJ23958@pathway.suse.cz

BUG=chrome-os-partner:51198
TEST=sudo -i;
     echo 10 > /proc/sys/kernel/hung_task_timeout_secs;
     echo HUNG_TASK > /sys/kernel/debug/provoke-crash/DIRECT &

Reviewed-by: Petr Mladek <pmladek@suse.cz>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
(cherry picked from commit 74f06bb72347302a19aac087314388ebd0e4fee9
 git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git)
Signed-off-by: Harry Pan <harry.pan@intel.com>

Change-Id: Ib446935061a63e25fbb49255559360ee1beb65c9
Reviewed-on: https://chromium-review.googlesource.com/337894
Commit-Ready: Harry Pan <harry.pan@intel.com>
Tested-by: Zhuo-hao Lee <zhuo-hao.lee@intel.com>
Reviewed-by: Benson Leung <bleung@chromium.org>