UPSTREAM: perf/x86: Add 'index' param to get_event_constraint() callback
This patch adds an index parameter to the get_event_constraint()
x86_pmu callback. It is expected to represent the index of the
event in the cpuc->event_list[] array. When the callback is used
for fake_cpuc (evnet validation), then the index must be -1. The
motivation for passing the index is to use it to index into another
cpuc array.

BUG=b:33115527
TEST=Verified that it compiles

Signed-off-by: Stephane Eranian <eranian@google.com>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Cc: bp@alien8.de
Cc: jolsa@redhat.com
Cc: kan.liang@intel.com
Cc: maria.n.dimakopoulou@gmail.com
Link: http://lkml.kernel.org/r/1416251225-17721-5-git-send-email-eranian@google.com
Signed-off-by: Ingo Molnar <mingo@kernel.org>
(cherry picked from commit 79cba822443a168c8f7f5b853d9c7225a6d5415e)
Signed-off-by: Gabriel Marin <gmx@chromium.org>

Change-Id: I31ce59a72678d7d2cd3387068139944d53d9e1ab
Reviewed-on: https://chromium-review.googlesource.com/417206
Commit-Ready: Gabriel Marin <gmx@chromium.org>
Tested-by: Gabriel Marin <gmx@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>