BACKPORT: perf/x86/intel: Make the HT bug workaround conditional on HT enabled
This patch disables the PMU HT bug when Hyperthreading (HT)
is disabled. We cannot do this test immediately when perf_events
is initialized. We need to wait until the topology information
is setup properly. As such, we register a later initcall, check
the topology and potentially disable the workaround. To do this,
we need to ensure there is no user of the PMU. At this point of
the boot, the only user is the NMI watchdog, thus we disable
it during the switch and re-enable it right after.

Having the workaround disabled when it is not needed provides
some benefits by limiting the overhead is time and space.
The workaround still ensures correct scheduling of the corrupting
memory events (0xd0, 0xd1, 0xd2) when HT is off. Those events
can only be measured on counters 0-3. Something else the current
kernel did not handle correctly.

BUG=b:33115527
TEST=Verified that it compiles

Signed-off-by: Stephane Eranian <eranian@google.com>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Cc: bp@alien8.de
Cc: jolsa@redhat.com
Cc: kan.liang@intel.com
Cc: maria.n.dimakopoulou@gmail.com
Link: http://lkml.kernel.org/r/1416251225-17721-13-git-send-email-eranian@google.com
Signed-off-by: Ingo Molnar <mingo@kernel.org>
(cherry picked from commit b37609c30e41264c4df4acff78abfc894499a49b)
Signed-off-by: Gabriel Marin <gmx@chromium.org>

Conflicts:
	arch/x86/kernel/cpu/perf_event.h
	arch/x86/kernel/cpu/perf_event_intel.c

Additional fixes:
  In perf_event.h and perf_event_intel.c:
  - replace PMU_FL_EXCL_ENABLED with ERF_EXCL_ENABLED
  - replace x86_pmu.flags with x86_pmu.er_flags

Change-Id: I73cc5d40900e5f082a566ec734b418ff405c10cd
Reviewed-on: https://chromium-review.googlesource.com/417211
Commit-Ready: Gabriel Marin <gmx@chromium.org>
Tested-by: Gabriel Marin <gmx@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>