UPSTREAM: netns: add rtnl cmd to add and get peer netns ids
With this patch, a user can define an id for a peer netns by providing a FD or a
PID. These ids are local to the netns where it is added (ie valid only into this
netns).

The main function (ie the one exported to other module), peernet2id(), allows to
get the id of a peer netns. If no id has been assigned by the user, this
function allocates one.

These ids will be used in netlink messages to point to a peer netns, for example
in case of a x-netns interface.

BUG=none
TEST="ip netns add foo" works again with selinux enabled.

Signed-off-by: Nicolas Dichtel <nicolas.dichtel@6wind.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Dylan Reid <dgreid@chromium.org>
(cherry picked from commit 0c7aecd4bde4b7302cd41986d3a29e4f0b0ed218)
Change-Id: Ibfcc9e95713144560645755e2bd83b6fbc276dfd
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/360915
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>