UPSTREAM: VFS: Introduce inode-getting helpers for layered/unioned fs environments
Introduce some function for getting the inode (and also the dentry) in an
environment where layered/unioned filesystems are in operation.

The problem is that we have places where we need *both* the union dentry and
the lower source or workspace inode or dentry available, but we can only have
a handle on one of them.  Therefore we need to derive the handle to the other
from that.

The idea is to introduce an extra field in struct dentry that allows the union
dentry to refer to and pin the lower dentry.

Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>

BUG=b:29259708
TEST=Built and booted on cyan

(cherry picked from commit 155e35d4daa804582f75acaa2c74ec797a89c615)
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Change-Id: I41dd35a07182b86f5d550d82703bbf01937fb3e6
Reviewed-on: https://chromium-review.googlesource.com/353780
Reviewed-by: Dylan Reid <dgreid@chromium.org>