UPSTREAM: r8152: restore hw settings
There is a capability which let the hw could change the settings
automatically when the power change to ON. However, the USB reset
would reset the settings to the hw default, so the driver has to
restore the relative settings. Otherwise, it would influence the
functions of the hw, and the compatibility for the USB hub and
USB host controller.

The relative settings are as following.
 - set the power down scale to 96.
 - enable the power saving function of USB 2.0.
 - disable the ALDPS of ECM mode.
 - set burst mode depending on the burst size.
 - enable the flow control of endpoint full.
 - set fifo empty boundary to 32448 bytes.
 - enable the function of exiting LPM when Rx OK occurs.
 - set the connect timer to 1.

Signed-off-by: Hayes Wang <hayeswang@realtek.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 65bab84c79077f3acbf7dcf2aec0588226d43663)

BUG=chrome-os-partner:53802
TEST=built/installed on gale and sentry. iperf no longer TX timeouts.

Change-Id: If8d4c44ee401a8d3b02c9a45f82b388a67da90cb
Signed-off-by: Grant Grundler <grundler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/348915
Commit-Ready: Zhifeng Cai <caiz@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>