UPSTREAM: arm64: pmu: add support for interrupt-affinity property
Historically, the PMU devicetree bindings have expected SPIs to be
listed in order of *logical* CPU number. This is problematic for
bootloaders, especially when the boot CPU (logical ID 0) isn't listed
first in the devicetree.

This patch adds a new optional property, interrupt-affinity, to the
PMU node which allows the interrupt affinity to be described using
a list of phandled to CPU nodes, with each entry in the list
corresponding to the SPI at the same index in the interrupts property.

Cc: Mark Rutland <mark.rutland@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
(cherry picked from commit d5efd9cc9cf2e422d064c912c7d5d985f52c1b2c)

BUG=chrome-os-partner:54420
TEST=Run `perf -ad sleep 3` on my oak and make sure the counters
     values are not corrupted.

Change-Id: I311b8610ec636fa1754631e5cd14d7a3b2744e9f
Signed-off-by: Ricky Liang <jcliang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/353222
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>