UPSTREAM: HID: logitech-hidpp: add a module parameter to keep firmware gestures
The Logitech T650 used to report 3 fingers swipes to the up as a press on the
Super key. When we switched the touchpad to the raw mode, we also disable such
firmware gesture and some users may rely on it.

Unfortunately, 3 finger swipes are still not supported in most of the Linux
environments, which means that we disabled a feature of the touchpad.

Allow users to revert the raw reporting mode and keep going with the firmware
gestures by providing a new module parameter.

BUG=chromium:617976
TEST=build

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit 9188dbaed68a4b23dc96eba165265c08caa7dc2a)
Signed-off-by: Adam Goode <agoode@chromium.org>

Change-Id: I547a7c9da954cd56620be3cef25e6010dcec1145
Reviewed-on: https://chromium-review.googlesource.com/361126
Commit-Ready: Adam Goode <agoode@chromium.org>
Tested-by: Adam Goode <agoode@chromium.org>
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>