UPSTREAM: drm/atomic: Clear crtcs, connectors and planes when clearing state
Users of the atomic state assume that if the pointer to a crtc, plane or
connector is not NULL in the respective object vector, than the state
for that object in *_states vector also won't be NULL. That assumption
was broken by drm_atomic_state_clear(), which would clear the state
pointer but leave the pointer to the object still set.

This fixes a NULL pointer dereference in i915 caused by the use of
drm_atomic_state_clear().

Cc: dri-devel@lists.freedesktop.org
Signed-off-by: Ander Conselvan de Oliveira <ander.conselvan.de.oliveira@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
(cherry picked from commit 8a5c0bde54ca0354ae8b7e9d37d17bc514aa6c42)
Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:54665
TEST=boot to ui on elm
TEST=HDMI work normal (plug/unplug/replug, display normal, show/hide cursor)
TEST=backlight on/off
TEST=suspend resume stress test finished 10000 iterations.
TEST=chameleon display_HotPlugAtSuspend.extended_unplugged
TEST=chameleon display_HotPlugAtSuspend.mirrored_unplugged

Change-Id: I0a2738bd468cedfa25791a9a80ae7d9ba44fd8a2
Reviewed-on: https://chromium-review.googlesource.com/362233
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>