UPSTREAM: watchdog: watchdog_core: Add watchdog registration deferral mechanism
Currently, watchdog subsystem require the misc subsystem to
register a watchdog. This may not be the case in case of an
early registration of a watchdog, which can be required when
the watchdog cannot be disabled.

This patch introduces a deferral mechanism to remove this requirement.

Signed-off-by: Jean-Baptiste Theou <jtheou@adeneo-embedded.us>
Reviewed-by: Guenter Roeck <linux@roeck-us.net>
Signed-off-by: Wim Van Sebroeck <wim@iguana.be>
(cherry picked from commit ef90174f821041313d42d99c1c8b35a3af64a910)

BUG=chrome-os-partner:49255
TEST=watchdog is resetting the device after timeout

Change-Id: I508d6acd73a7a873b1102f597f651b9fee571a66
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/334852
Commit-Ready: Dmitry Torokhov <dtor@chromium.org>
Tested-by: Matthew McClintock <mmcclint@qca.qualcomm.com>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>