UPSTREAM: mac80211: mac80211: Check SN for deactivated mpaths
When processing a PREQ or PREP it's critical to use the incoming SN. If
that is improperly done routing loops and other types of badness can
happen. But the code was always processing path messages for deactivated
paths. This path fixes that so that if we have a valid SN then we use it
to verify that it is a message we can accept. For reference the relevant
section of the standard is 13.10.8.4 which doesn't address the deactivated
path case at all.

I also included a special case for when our peer reboots or restarts
networking. This is an important case because without it there can be a
very long delay before we accept path messages from that peer. It's also a
simple case and intimately associated with processing messages for
deactivated paths so I used one patch instead of two.

(cherry picked from commit "d8f0300a7aca5cd9208112104c64d894ad82da1f")
(source: git://git.kernel.org/pub/scm/linux/kernel/git/kvalo/ath.git)

BUG=chrome-os-partner:49352
TEST=build/boot on Arkham and verified iperf throughput.

Signed-off-by: Alexis Green <agreen@cococorp.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/322375
Commit-Ready: Srinivasa duvvuri <sduvvuri@chromium.org>
Tested-by: Peter Oh <poh@qca.qualcomm.com>
Tested-by: Srinivasa duvvuri <sduvvuri@chromium.org>
Reviewed-by: Julan Hsu <julanhsu@google.com>
Reviewed-by: Srinivasa duvvuri <sduvvuri@chromium.org>

Change-Id: I55dde29a634e899c37641379da7cc35aa489e4cb
Signed-off-by: Srinivasa duvvuri <sduvvuri@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/340249
Commit-Ready: Grant Grundler <grundler@chromium.org>
Tested-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>